Add "CodeQL: Trim Cache" using new mode parameter of `evaluation/clearCache`
The purpose of this change is to add a command to clear the cache
except for `cached` predicates. This is the
same behaviour as `codeql database cleanup --mode=normal`.
In contrast, the existing "CodeQL: Clear Cache" corresponds to `--mode=brutal`.

This change depends on adding a new `mode` parameter to the
`evaluation/clearCache` method of the query server. This is the same
method that the existing "CodeQL: Clear Cache" command targets, which means
that this solution will not fail when used with older query servers
but will have the same behaviour as the existing command.
This somewhat decouples the release process of the VSCode extension and
the CodeQL release.

Testing done so far: manual (VSCode extension debug run). I haven't seen
automated tests for the other command either.

As part of adding the new "mode" parameter to the clearCache method, the
official query server spec will need to be updated to match `new-messages.ts`
here.