Remove fallback behavior for database unbundle
The CodeQL CLI always supports the `database unbundle` command since
https://github.com/github/vscode-codeql/commit/140d369098b286a006322432777b88649832c954
so we can remove the fallback behavior.

There were some places which were not passing in the CodeQL CLI server,
but these always have access to the CLI server, so this just passes them
in.

The only change in behavior (in terms of the fallback behavior) is in
the `new-query.test.ts` test.