Avoid unhandled errors in build streams
The .pipe() method on Node.js streams have a gotcha in that they
do not propagate errors downstream. In this case this meant that if
compilation failed with an error, it would not be visible to Gulp
because Gulp only sees the stream that writes to a file, not the
intermediate 'esbuild' stream that generated the error.

This changes a few uses of .pipe() to the more modern stream.pipeline(),
which propagates errors correctly.