Fix join-order scoring for recursive predicates
The evaluator log summary JSON needs to be read in two passes, to avoid
trying to calculate scores for a recursive layer before all the events
for it have been processed. Doing it in one pass meant that, for
predicates associated with an IN_LAYER event that came after the
corresponding COMPUTE_RECURSIVE event, we were missing their delta sizes
and were not even attempting to their join-order scores.

I had to remove some of the generic machinery for registering log
processors, but the JoinOrderScanner was the only processor being
registered anyway.